test(scoreboard): Scoreboard 도메인 테스트 작성
테스트 내용
- `LottoServiceTest` 내 당첨 등수 계산 로직에 대한 테스트를 `ScoreboardTest`로 이동
- `LottoServiceTest` 테스트 삭제
